All values in header now converted to bytes. HeaderGenerator returns full header byte array
#story[741] #pair[jth102, sbe67]

+++ racevision/test_mock/src/main/java/seng302/team18/test_mock/connection/Server.java
@@ -90,7 +90,9 @@ public class Server {
     public void openServer() {
         try {
             serverSocket = new ServerSocket(PORT);
-            HeaderGenerator.generateHeader((byte)2, 6); //TODO remove this line if you see it, sam forgot to delete it
+
+            HeaderGenerator.generateHeader(2, (short) 6); //TODO remove this line if you see it, sam forgot to delete it
+
         } catch (IOException e) {
             System.err.println("Could not listen on port " + PORT);
             System.err.println("Exiting program");
